Removed UnitTests project - it should be refactored and added back from the test-framework branch in the near future.
Everything below is technically obsolete and reverted in later commit for easier PR merge:

Removed old, static Karma class
Started using new karma retrieval throughout the project.
Also found a bunch of queries done from the views (!) and static dependencies which were commented out for the time being with a TODO message.
