+++ Whoaverse/Voat.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Voat", "Whoaverse\Voat.csproj", "{1D1F318C-EB40-4E3B-96EB-E9AB0D66DE30}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{218EEACE-623C-4D8C-A87A-E4AD4DDD5EAD}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{1D1F318C-EB40-4E3B-96EB-E9AB0D66DE30}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D1F318C-EB40-4E3B-96EB-E9AB0D66DE30}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1D1F318C-EB40-4E3B-96EB-E9AB0D66DE30}.Release|Any CPU.Build.0 = Release|Any CPU
-		{218EEACE-623C-4D8C-A87A-E4AD4DDD5EAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{218EEACE-623C-4D8C-A87A-E4AD4DDD5EAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{218EEACE-623C-4D8C-A87A-E4AD4DDD5EAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{218EEACE-623C-4D8C-A87A-E4AD4DDD5EAD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
