+++ Whoaverse/Whoaverse/Controllers/HomeController.cs
@@ -30,6 +30,7 @@ using Voat.Utils.Components;
 
 namespace Voat.Controllers
 {
+    using Queries.Karma;
 
 
     public class HomeController : Controller
@@ -174,7 +175,7 @@ namespace Voat.Controllers
             }
 
             // verify recaptcha if user has less than 25 CCP
-            var userCcp = Karma.CommentKarma(User.Identity.Name);
+            var userCcp = await _db.GetCommentKarmaAsync(User.Identity.Name);
             if (userCcp < 25)
             {
                 bool isCaptchaCodeValid = await ReCaptchaUtility.Validate(Request);
@@ -191,7 +192,7 @@ namespace Voat.Controllers
             }
 
             // if user CCP or SCP is less than -50, allow only X submissions per 24 hours
-            var userScp = Karma.LinkKarma(User.Identity.Name);
+            var userScp = await _db.GetLinkKarmaAsync(User.Identity.Name);
             if (userCcp <= -50 || userScp <= -50)
             {
                 var quotaUsed = Utils.User.UserDailyPostingQuotaForNegativeScoreUsed(User.Identity.Name);
