+++ Whoaverse/Whoaverse/Controllers/MessagingController.cs
@@ -25,6 +25,8 @@ using Voat.Utils;
 
 namespace Voat.Controllers
 {
+    using Queries.Karma;
+
     public class MessagingController : Controller
     {
         private readonly whoaverseEntities _db = new whoaverseEntities();
@@ -275,8 +277,8 @@ namespace Voat.Controllers
         {
             if (!ModelState.IsValid) return View();
             if (privateMessage.Recipient == null || privateMessage.Subject == null || privateMessage.Body == null) return RedirectToAction("Sent", "Messaging");
-
-            if (Karma.CommentKarma(User.Identity.Name) < 100)
+            var commentKarma = await _db.GetCommentKarmaAsync(User.Identity.Name);
+            if (commentKarma < 100)
             {
                 bool isCaptchaValid = await ReCaptchaUtility.Validate(Request);
