+++ Whoaverse/Whoaverse/Controllers/SubversesController.cs
@@ -28,6 +28,8 @@ using Voat.Utils;
 
 namespace Voat.Controllers
 {
+    using Queries.Karma;
+
     public class SubversesController : Controller
     {
         private readonly whoaverseEntities _db = new whoaverseEntities();
@@ -126,9 +128,9 @@ namespace Voat.Controllers
 
             int minimumCcp = Convert.ToInt32(ConfigurationManager.AppSettings["minimumCcp"]);
             int maximumOwnedSubs = Convert.ToInt32(ConfigurationManager.AppSettings["maximumOwnedSubs"]);
-
+            var commentKarma = await _db.GetCommentKarmaAsync(User.Identity.Name);
             // verify recaptcha if user has less than minimum required CCP
-            if (Karma.CommentKarma(User.Identity.Name) < minimumCcp)
+            if (commentKarma < minimumCcp)
             {
                 // begin recaptcha check
                 bool isCaptchaCodeValid = await ReCaptchaUtility.Validate(Request);
