+++ Whoaverse/Whoaverse/Controllers/WebApiController.cs
@@ -23,6 +23,9 @@ using Voat.Utils;
 
 namespace Voat.Controllers
 {
+    using System.Threading.Tasks;
+    using Queries.Karma;
+
     public class WebApiController : ApiController
     {
         private readonly whoaverseEntities _db = new whoaverseEntities();
@@ -305,7 +308,7 @@ namespace Voat.Controllers
         /// </summary>
         /// <param name="userName">The username for which to fetch basic information.</param>
         [HttpGet]
-        public ApiUserInfo UserInfo(string userName)
+        public async Task<ApiUserInfo> UserInfo(string userName)
         {
             if (userName != "deleted" && !Utils.User.UserExists(userName))
             {
@@ -323,8 +326,8 @@ namespace Voat.Controllers
             var resultBadgesList = userBadgesList.Select(item => new ApiUserBadge {Awarded = item.Awarded, BadgeName = item.Badge.BadgeName}).ToList();
 
             resultModel.Name = userName;
-            resultModel.CCP = Karma.CommentKarma(userName);
-            resultModel.LCP = Karma.LinkKarma(userName);
+            resultModel.CCP = await _db.GetCommentKarmaAsync(userName);
+            resultModel.LCP = await _db.GetLinkKarmaAsync(userName);
             resultModel.RegistrationDate = Utils.User.GetUserRegistrationDateTime(userName);
             resultModel.Badges = resultBadgesList;
