+++ Whoaverse/Whoaverse/Views/Home/comments.cshtml
@@ -19,7 +19,8 @@
     int commentContributionPoints = 0;
     if (User.Identity.IsAuthenticated)
     {
-        commentContributionPoints = Karma.CommentKarma(User.Identity.Name);
+        //TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.
+        commentContributionPoints = 0;//Karma.CommentKarma(User.Identity.Name);
     }
 }
