+++ Whoaverse/Whoaverse/Views/Messaging/Compose.cshtml
@@ -59,8 +59,9 @@
                                 @Html.ValidationMessageFor(model => model.Body)
                             </div>
                         </div>
-
-                        @if (Karma.CommentKarma(User.Identity.Name) < 100)
+                        @*TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.*@
+                        @*@if (Karma.CommentKarma(User.Identity.Name) < 100)*@
+                        @if(false)
                         {
                             <div class="panel panel-default">
                                 <div class="panel-heading">Are you a human?</div>
