+++ Whoaverse/Whoaverse/Views/Search/Index.cshtml
@@ -34,7 +34,8 @@
     int commentContributionPoints = 0;
     if (User.Identity.IsAuthenticated)
     {
-        commentContributionPoints = Karma.CommentKarma(User.Identity.Name);
+        //TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.
+        commentContributionPoints = 0; //Karma.CommentKarma(User.Identity.Name);
     }
 }
