+++ Whoaverse/Whoaverse/Views/Sets/_SingleSetPage.cshtml
@@ -18,7 +18,8 @@
 
     if (User.Identity.IsAuthenticated)
     {
-        commentContributionPoints = Karma.CommentKarma(User.Identity.Name);
+        //TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.
+        commentContributionPoints = 0; //Karma.CommentKarma(User.Identity.Name);
     }
 }
