+++ Whoaverse/Whoaverse/Views/Shared/Sidebars/_Sidebar.cshtml
@@ -113,7 +113,8 @@
                             Minimum CCP required to downvote in this sub: @Model.minimumdownvoteccp
                             @if (User.Identity.IsAuthenticated)
                             {
-                            <p>Your CCP for this sub: @Karma.CommentKarmaForSubverse(User.Identity.Name, Model.name)</p>
+                                /*TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.*/
+                            <p>Your CCP for this sub: 0 @*@Karma.CommentKarmaForSubverse(User.Identity.Name, Model.name)*@</p>
                             }
                             </p>
                         </div>
