+++ Whoaverse/Whoaverse/Views/Shared/_AddSubmissionLinkpost.cshtml
@@ -91,7 +91,9 @@
             </div>
         </div>
 
-        @if (Karma.CommentKarma(User.Identity.Name) < 25)
+        @*TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.*@
+        @*@if (Karma.CommentKarma(User.Identity.Name) < 25)*@
+        @if(false)
         {
             <div class="alert alert-info">Are you a human? Please solve the reCAPTCHA challenge below (sorry, this is to prevent spam):</div>
