+++ Whoaverse/Whoaverse/Views/Shared/_Layout.cshtml
@@ -246,7 +246,8 @@
                         <p>Welcome to @ConfigurationManager.AppSettings["siteName"]!</p>
                         <p>In order to downvote comments or submissions you need to have at least 100 comment contribution points.</p>
                         <p>Every time someone upvotes one of your comments, you gain 1 comment contribution point.</p>
-                        <p>You currently have @Karma.CommentKarma(User.Identity.Name) comment contribution points.</p>
+                        @*//TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.*@
+                        @*<p>You currently have @Karma.CommentKarma(User.Identity.Name) comment contribution points.</p>*@
                         <p>Tip: a good way to earn more contribution points is to take part in discussions or say hi in /v/introductions.</p>
                         <p>This requirement was established as a measure to encourage new members to get involved and participate in discussions.</p>
                     </div>
@@ -269,7 +270,8 @@
                         <p>In order to be able to upvote comments or submissions without limitations, you need to have at least 20 comment contribution points.</p>
                         <p>You are currently limited to 10 upvotes per 24 hours.</p>
                         <p>Every time someone upvotes one of your comments, you gain 1 comment contribution point.</p>
-                        <p>You currently have @Karma.CommentKarma(User.Identity.Name) comment contribution points.</p>
+                        @*TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.*@
+                        @*<p>You currently have @Karma.CommentKarma(User.Identity.Name) comment contribution points.</p>*@
                         <p>Tip: a good way to earn more contribution points is to take part in discussions or say hi in /v/introductions.</p>
                         <p>This requirement was established as a measure to encourage new members to get involved and participate in discussions.</p>
                     </div>
