+++ Whoaverse/Whoaverse/Views/Subverses/SubverseIndex.cshtml
@@ -21,7 +21,8 @@
     int commentContributionPoints = 0;
     if (User.Identity.IsAuthenticated)
     {
-        commentContributionPoints = Karma.CommentKarma(User.Identity.Name);
+        //TODO: DB call in a view? That's a no-no. Appropriate property should be in the model.
+        commentContributionPoints = 0;/*Karma.CommentKarma(User.Identity.Name);*/
     }
 
     switch (sortingMode)
